requirements.txt
boto
docopt
python-crontab